refactor: integration tests separated and with external api, optional path for hot db reset
